move `apply_to_items` to the custom dicttoolz
