+++ ceos_alos2/dicttoolz.py
@@ -1,4 +1,5 @@
 from tlz.itertoolz import concat, groupby
+from tlz.itertoolz import identity as passthrough
 
 from ceos_alos2.utils import unique
 
@@ -24,3 +25,7 @@ def zip_default(*mappings, default=None):
     all_keys = unique(concat(map(list, mappings)))
 
     return {k: [m.get(k, default) for m in mappings] for k in all_keys}
+
+
+def apply_to_items(funcs, mapping, default=passthrough):
+    return {k: funcs.get(k, default)(v) for k, v in mapping.items()}
